Added developer.py and upload.py to the hidden folder
These two files add a nenw vulnerability, namely, a file upload vulnerability. This vulnerability
is detectable via w3af, and when used, gives access to the user to a root shell.  I'm thinking we
use the w3af to initially detect the vulnerability, and then tell them to insert a file that can
be used for hijacking at first.  Then, we tell them how to use w3af to do this.

I'm still debugging this, so it may not completely work yet.